pyproject.toml
[project]
name = "agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "debugpy>=1.8.17",
    "ipython>=9.5.0",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-postgres>=2.0.23",
    "mcp>=1.14.1",
    "numpy>=2.3.3",
    "psycopg2>=2.9.10",
    "psycopg[binary,pool]>=3.2.10",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.15.0",
    "you-get>=0.4.1743",
]
